# public/openapi.yaml
openapi: 3.1.0
info:
  title: DBackup API
  version: 2.4.0
  description: |
    REST API for DBackup - a self-hosted database backup automation platform with encryption, compression, and smart retention.

    ## Authentication

    DBackup supports two authentication methods:

    - **Session Authentication** - Used automatically when logged in via the web UI. Session cookies are sent with each request.
    - **API Key Authentication** - For scripts, CI/CD pipelines, and external integrations. Create an API key under **Access Management → API Keys**.

    > API keys do not inherit SuperAdmin privileges. Only explicitly assigned permissions are available.

    ## Rate Limits

    | Request Type | Default Limit |
    |---|---|
    | Authentication (login, etc.) | 5/min per IP |
    | GET requests | 100/min per IP |
    | POST / PUT / DELETE | 20/min per IP |

    Rate limits are configurable in Settings → Rate Limits.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
...
servers:
  - url: "{protocol}://{host}/api"
    description: DBackup Instance
    variables:
      protocol:
        default: https
        enum: [http, https]
      host:
        default: your-dbackup-instance.com
security:
  - BearerAuth: []
  - SessionAuth: []
...
paths:
...
  /adapters:
    get:
      tags: [Adapters]
      summary: List Adapters
      description: |
        List all adapter configurations of a given type. Config fields are automatically decrypted in the response.

        **Permission depends on type:**
        - `database` → `sources:read`
        - `storage` → `destinations:read`
        - `notification` → `notifications:read`
      operationId: listAdapters
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [database, storage, notification]
          description: Adapter category
      responses:
        "200":
          description: Adapter list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdapterConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Adapters]
      summary: Create Adapter
      description: |
        Create a new adapter configuration. Sensitive config fields (passwords, keys) are automatically encrypted before storage.

        **Permission depends on type:**
        - `database` → `sources:write`
        - `storage` → `destinations:write`
        - `notification` → `notifications:write`
      operationId: createAdapter
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdapterRequest"
      responses:
        "201":
          description: Adapter created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdapterConfig"
        "409":
          description: Adapter with this name and type already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API Key authentication. Create an API key under **Access Management → API Keys**.

        Format: `dbackup_xxxxxxxx`
    SessionAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie (browser login)
...
  responses:
    Unauthorized:
      description: Missing, invalid, disabled, or expired credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Valid credentials but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Human-readable error message"
...
    AdapterConfig:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: "Production MySQL"
        type:
          type: string
          enum: [database, storage, notification]
        adapterId:
          type: string
          description: Adapter type identifier
          example: mysql
        config:
          type: object
          description: Adapter-specific configuration (decrypted)
        metadata:
          type: [object, "null"]
          description: "Auto-detected metadata (version, status, etc.)"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAdapterRequest:
      type: object
      required: [name, type, adapterId, config]
      properties:
        name:
          type: string
          example: "Production MySQL"
        type:
          type: string
          enum: [database, storage, notification]
        adapterId:
          type: string
          description: |
            Adapter type identifier. Available values:

            **Database:** `mysql`, `mariadb`, `postgres`, `mongodb`, `sqlite`, `mssql`, `redis`

            **Storage:** `local-filesystem`, `s3-aws`, `s3-generic`, `s3-r2`, `s3-hetzner`, `google-drive`, `dropbox`, `onedrive`, `sftp`, `ftp`, `webdav`, `smb`, `rsync`

            **Notification:** `discord`, `slack`, `teams`, `generic-webhook`, `gotify`, `ntfy`, `telegram`, `twilio-sms`, `email`
        config:
          type: object
          description: Adapter-specific configuration. See adapter documentation for fields.
...